+++ app/src/main/java/com/ceaver/assin/trades/input/TradeInputViewModel.kt
@@ -11,6 +11,7 @@ import com.ceaver.assin.threading.BackgroundThreadExecutor
 import com.ceaver.assin.trades.Trade
 import com.ceaver.assin.trades.TradeRepository
 import com.ceaver.assin.trades.TradeType
+import java.math.BigDecimal
 import java.time.LocalDate
 
 class TradeInputViewModel : ViewModel() {
@@ -43,15 +44,15 @@ class TradeInputViewModel : ViewModel() {
         TradeRepository.saveTradeAsync(trade, true) { status.value = TradeInputStatus.END_SAVE }
     }
 
-    fun onSaveTradeClick(buySymbol: Title, buyAmount: Double, sellSymbol: Title, sellAmount: Double, tradeDate: LocalDate, comment: String?) {
+    fun onSaveTradeClick(buySymbol: Title, buyAmount: BigDecimal, sellSymbol: Title, sellAmount: BigDecimal, tradeDate: LocalDate, comment: String?) {
         saveTrade(trade.value!!.copy(buyTitle = buySymbol, buyAmount = buyAmount, sellTitle = sellSymbol, sellAmount = sellAmount, tradeDate = tradeDate, comment = comment))
     }
 
-    fun onDepositClick(buySymbol: Title, buyAmount: Double, tradeDate: LocalDate, comment: String?) {
+    fun onDepositClick(buySymbol: Title, buyAmount: BigDecimal, tradeDate: LocalDate, comment: String?) {
         saveTrade(trade.value!!.copy(buyTitle = buySymbol, buyAmount = buyAmount, tradeDate = tradeDate, comment = comment))
     }
 
-    fun onWithdrawClick(sellSymbol: Title, sellAmount: Double, tradeDate: LocalDate, comment: String?) {
+    fun onWithdrawClick(sellSymbol: Title, sellAmount: BigDecimal, tradeDate: LocalDate, comment: String?) {
         saveTrade(trade.value!!.copy(sellTitle = sellSymbol, sellAmount = sellAmount, tradeDate = tradeDate, comment = comment))
     }
